chore: update variant files with additional services and minor refactorings
- `app.component.ts`: Added `MessageService` import to enhance component functionality.
- `companies.component.ts`: Commented out unused code in `onRowEditSave` method.
- `schedules.component.ts`: Updated service import path for consistency; no functional changes.
- `base.service.ts`: Added a `patch` method to support partial updates.
- Added new DTO files: `delete.res.dto.ts`, `image.res.dto.ts`, and `role.res.dto.ts`.

Refactored imports for better readability and maintainability.
