+++ .vscode/settings.json
@@ -5,9 +5,11 @@
         "Budi",
         "devkit",
         "Dule",
+        "fileupload",
         "icondisplay",
         "inputtext",
         "Lita",
+        "messageservice",
         "primeflex",
         "primeicons",
         "primeng",
