+++ src/app/models/dto/delete.res.dto.ts
@@ -1,0 +1,3 @@
+export interface DeleteResDto {
+    message: string
+}
