+++ src/app/models/dto/file/image.res.dto.ts
@@ -1,0 +1,3 @@
+export interface ImageResDto{
+    url: string;
+}
