+++ src/app/models/dto/role/role.res.dto.ts
@@ -1,0 +1,5 @@
+export interface RoleResDto {
+    id: string;
+    roleCode: string;
+    roleName: string;
+}
