+++ src/app/services/payroll/payroll.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from "@angular/core";
-import { PayrollsResDto } from "../models/dto/payrolls/payrolls.res.dto";
-import { BaseService } from "./base/base.service";
+import { PayrollsResDto } from "../../models/dto/payrolls/payrolls.res.dto";
+import { BaseService } from "../base/base.service";
 
 @Injectable({
   providedIn: "root"
